:recycle: Partial refactor
