+++ src/contexts/Counter/application/kick/CounterKicker.py
