+++ src/contexts/Counter/domain/entities/Counter.py
