+++ src/contexts/Counter/domain/entities/User.py
