+++ src/contexts/Counter/domain/repositories/CounterRepository.py
