+++ src/contexts/Counter/domain/repositories/UserRepository.py
