+++ src/contexts/Counter/domain/services/counter.py
