+++ src/contexts/Counter/domain/value_objects/CounterId.py
@@ -1,0 +1,8 @@
+from uuid import UUID
+from dataclasses import dataclass
+
+from ..properties import UuidProperty
+
+
+class CounterId(UuidProperty):
+    pass
