+++ src/contexts/Counter/domain/value_objects/CounterPrivate.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+from ..properties import BooleanProperty
+
+
+class CounterPrivate(BooleanProperty):
+    pass
