+++ src/contexts/Counter/domain/value_objects/CounterStatus.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from ..properties import IntegerProperty
+
+
+class CounterStatus(IntegerProperty):
+    pass
+
