+++ src/contexts/Counter/domain/value_objects/UserId.py
@@ -1,0 +1,7 @@
+from uuid import UUID
+from dataclasses import dataclass
+
+from ..properties import UuidProperty
+
+class UserId(UuidProperty):
+    pass
