+++ tests/Counter/infrastructure/FakeCounterRepository.py
@@ -3,15 +3,15 @@ from typing import (
     Optional
 )
 
-from src.Counter.domain.entities import (
+from src.contexts.Counter.domain.entities import (
     Counter
 )
 
-from src.Counter.domain.value_objects import (
+from src.contexts.Counter.domain.value_objects import (
     CounterId
 )
 
-from src.Counter.domain.repositories import CounterRepository
+from src.contexts.Counter.domain.repositories import CounterRepository
 
 
 class FakeCounterRepository(CounterRepository):
